backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# AI/ML
numpy==1.24.3
Cython==3.0.5

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0

# Development
pytest==7.4.3
pytest-django==4.7.0
